Rename project files to kebab-case
