+++ config/cards.js
